Download questionnaire results

Download results for a particular typeform questionnaire.

Usage

get_responses(
  form_id,
  api = NULL,
  page_size = 25,
  since = NULL,
  until = NULL,
  after = NULL,
  before = NULL,
  completed = NULL,
  query = NULL,
  fields = NULL
)

Arguments

form_id

The form id of the typeform you want the results for.

api

Default NULL. Your private api key. If api is NULL, the environment variable Sys.getenv("typeform_api2") is used.

page_size

Maximum number of responses. Default value is 25. Maximum value is 1000.

since

default NULL. Fetch only the results after a specific date and time. If NULL return all results. This should be a date time object. The timezone of the object will be converted to UTC.

until

default NULL. Similar to since.

after

default NULL. Fetch only the results after a specific date and time. If NULL return all results. If you use the after parameter, the responses will be sorted in the order that our system processed them (instead of the default order, submitted_at). This ensures that you can traverse the complete set of responses without repeating entries.

before

default NULL. Similar to after

completed

default NULL, return all results. Fetch only completed results (TRUE), or only not-completed results (=FALSE). If NULL return all results. Warning. It's not possible to determine completed/non-completed results.

query

Limit request to only responses that that include the specified term.

fields

Not implemented. Pull requests welcome

Value

A list. The first value is meta information. Subsequent elements are questions..

See also